4月30号提交

+++ bupt-hotel-2/src/main/java/com/example/dao/entity/DetailedFeesDO.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 空调费用详表，旅客使用空调，发送一个请求【开关机、调温、调风速】，就结算上一次的请求。一条记录以空调编号、开始时间作为唯一标识。
  */
@@ -45,6 +43,6 @@ public class DetailedFeesDO {
     /*
     该段时间费用
      */
-    private BigDecimal fee;
+    private Double fee;
 
 }
