+++ bupt-hotel-2/src/main/java/com/example/dto/req/AdminStartReqDTO.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.HashMap;
 
 @Data
@@ -29,5 +28,5 @@ public class AdminStartReqDTO {
     /*
     计价规则：每档风速+每度温度：分钟价格
      */
-    private HashMap<String, BigDecimal> priceMap;
+    private HashMap<String, Double> priceMap;
 }
