+++ bupt-hotel-2/src/main/java/com/example/dto/resp/FrontendSettleBillRespDTO.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,21 +12,21 @@ public class FrontendSettleBillRespDTO {
     /*
     房间单价
      */
-    private BigDecimal roomPrice;
+    private Double roomPrice;
     /*
     住宿天数
      */
-    private BigDecimal days;
+    private Double days;
     /*
     房费
      */
-    private BigDecimal roomFee;
+    private Double roomFee;
     /*
     空调费
      */
-    private BigDecimal airConditionerFee;
+    private Double airConditionerFee;
     /*
     总费用
      */
-    private BigDecimal totalFee;
+    private Double totalFee;
 }
