+++ bupt-hotel-frontend/src/main/java/com/example/dto/resp/SettleBillRespDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+import java.math.Double;
 
 @Data
 @AllArgsConstructor
@@ -14,21 +14,21 @@ public class SettleBillRespDTO {
     /*
     房间单价
      */
-    private BigDecimal roomPrice;
+    private Double roomPrice;
     /*
     住宿天数
      */
-    private BigDecimal days;
+    private Double days;
     /*
     房费
      */
-    private BigDecimal roomFee;
+    private Double roomFee;
     /*
     空调费
      */
-    private BigDecimal airConditionerFee;
+    private Double airConditionerFee;
     /*
     总费用
      */
-    private BigDecimal totalFee;
+    private Double totalFee;
 }
